requirements.txt
fastapi
uvicorn
numpy
pillow
tensorflow-cpu